added indexes to folders and refactored imports/exports, added apiClient filee as the only source of axios and generic req functions
