+++ src/client/src/api/index.ts
@@ -1,0 +1,2 @@
+export * as notesAPI from './notesAPI';
+export * from './apiClient';
