+++ src/client/src/features/notes/NoteCard.tsx
@@ -1,9 +1,10 @@
-import { CardActionArea, makeStyles, Theme, Typography } from '@material-ui/core';
+import { makeStyles, Theme, Typography } from '@material-ui/core';
 import Menu from '@material-ui/core/Menu';
 import MenuItem from '@material-ui/core/MenuItem';
 import ListItemIcon from '@material-ui/core/ListItemIcon';
 import IconButton from '@material-ui/core/IconButton';
 import Card from '@material-ui/core/Card';
+import CardActionArea from '@material-ui/core/CardActionArea';
 import CardHeader from '@material-ui/core/CardHeader';
 import CardContent from '@material-ui/core/CardContent';
 import MoreVertIcon from '@material-ui/icons/MoreVert';
@@ -28,7 +29,7 @@ import {
   SNACKBAR_NOTE_DELETE_SUCCESS,
   SNACKBAR_NOTE_PIN_ERROR,
   SNACKBAR_NOTE_PIN_SUCCESS,
-} from 'constants/constants';
+} from 'constants/const';
 
 const useStyles = makeStyles((theme: Theme) => {
   return {
