+++ src/client/src/features/notes/index.ts
@@ -2,6 +2,7 @@ export { Notes } from './Notes';
 export { AddNote } from './AddNote';
 export { Note } from './Note';
 export { EditNote } from './EditNote';
+export { NoteCard } from './NoteCard';
 export {
   default as notesReducer,
   deleteNote,
