+++ src/client/src/mocks/index.ts
@@ -1,0 +1,3 @@
+export { notes } from './handlers';
+export { mockData, mockNote } from './mockData';
+export { rest, seedMockData, server } from './mockServer';
