+++ src/client/src/setupTests.ts
@@ -3,7 +3,7 @@
 // expect(element).toHaveTextContent(/react/i)
 // learn more: https://github.com/testing-library/jest-dom
 import '@testing-library/jest-dom';
-import { seedMockData, server } from 'mocks/mockServer';
+import { seedMockData, server } from 'mocks';
 
 beforeAll(() => {
   seedMockData();
